requirements-dbc.txt
appnope==0.1.4
asttokens==3.0.1
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
dab_project_dec @ file:///Users/saurabh/Desktop/databricks_asset_bundles/dab_project_dec/dist/dab_project_dec-0.0.1-py3-none-any.whl#sha256=a88c60b5c93d381ba2077a27cd5224840489d002f07d35025b399ba11936a7b5
databricks-connect==15.4.0
databricks-sdk==0.74.0
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14
